controle et utilisateur1 done

+++ Back/app/Http/Controllers/ControleController.php
@@ -12,16 +12,7 @@ class ControleController extends Controller
      */
     public function index()
     {
-        $controles = Controle::all(); // Récupérer tous les éléments de la table
-        return response()->json($controles);
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
+        return Controle::all();
     }
 
     /**
@@ -29,39 +20,20 @@ class ControleController extends Controller
      */
     public function store(Request $request)
     {
-        try
-        {
-        $test = Controle::where('nom', $request->nom)->where('code', $request->code)->first();
-        if ($test) {
-            return response()->json(['error' => 'Le controle a déjà été créé!']);
+        try{
+            $test = Controle::where('nom', $request->nom)->where('code', $request->code)->first();
+            if ($test) {
+                return response()->json(['error' => 'Le controle a déjà été créé!']);
+            }
+            Controle::create([
+                'nom' => $request->nom,
+                'code' => $request->code,
+            ]);
+            return response()->json(['message' => 'Le controle a été créé avec succès!']);
+
+        } catch (\Throwable $th) {
+            return response()->json(['error' => $th->getMessage()]);
         }
-
-        // Créer un nouvel élément dans la base de données
-        $controle = Controle::create([
-            'nom' => $request->nom,
-            'code' => $request->code,
-        ]);
-
-        return response()->json(['message' => 'Le controle a été créé avec succès!', 'controle' => $controle]);
-    } catch (\Throwable $th) {
-        return response()->json(['error' => $th->getMessage()]);
-    }
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(string $id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-        //
     }
 
     /**
@@ -70,25 +42,24 @@ class ControleController extends Controller
     public function update(Request $request, string $id)
     {
         try{
+            $controle = Controle::findOrFail($id);
 
-        
-        $controle = Controle::findOrFail($id); // Trouver l'élément
+            if ($controle->nom != $request->nom) {
+                $test = Controle::where('nom', $request->nom)->first();
+                if ($test) {
+                    return response()->json(['error' => 'Ce controle a déjà été créé!']);
+                }
+            }
 
-        $test = Controle::where('nom', $request->nom)->where('code', $request->code)->first();
-        if ($test) {
-            return response()->json(['error' => 'Ce controle a déjà été créé!']);
+            $controle->update([
+                'nom' => $request->nom,
+                'code' => $request->code
+            ]);
+            return response()->json(['message' => 'Le controle a été mis à jour avec succès!']);
+
+        } catch (\Throwable $th) {
+            return response()->json(['error' => $th->getMessage()]);
         }
-
-        // Mettre à jour l'élément
-        $controle->update([
-            'nom' => $request->nom,
-            'code' => $request->code
-        ]);
-
-        return response()->json(['message' => 'Le controle a été mis à jour avec succès!', 'controle' => $controle]);
-    } catch (\Throwable $th) {
-        return response()->json(['error' => $th->getMessage()]);
-    }
     }
 
     /**
@@ -96,8 +67,11 @@ class ControleController extends Controller
      */
     public function destroy(string $id)
     {
-        $controle = Controle::findOrFail($id); 
-        $controle->delete(); 
-        return response()->json(['message' => 'Le controle a été supprimé avec succès']); 
+        $controle = Controle::findOrFail($id);
+        if (!$controle) {
+            return response()->json(['error' => "Ce controle n'a pas été trouvé!"]);
+        }
+        $controle->delete();
+        return response()->json(['message' => 'Le controle a été supprimé avec succès']);
     }
 }
